Brdesk 5617

+++ resources/views/acomodacao/index.blade.php
@@ -24,7 +24,7 @@
         <a href="/acomodacao.edit.{{$obj->id}}"><i class="fa fa-edit"></i></a>
       </td>
       <td width="1%">
-        <a href="/acomodacao.delete.{{$obj->id}}"><i class="fa fa-trash"></i></a>
+        <a href="/acomodacao.delete.{{$obj->id}}" onclick="return confirm('Tem certeza de que deseja excluir este item?');"><i class="fa fa-trash"></i></a>
       </td>
       <td>
         {{$obj->descricao}}
