+++ resources/views/cidade/index.blade.php
@@ -33,7 +33,7 @@
                                 </td>
 
                                 <td width="1%">
-                                    <a href="/cidade.delete.{{$obj->id}}"><i class="fa fa-trash"></i></a>
+                                    <a href="/cidade.delete.{{$obj->id}}" onclick="return confirm('Tem certeza de que deseja excluir este item?');"><i class="fa fa-trash"></i></a>
                                 </td>
 
                                 <td>
