+++ resources/views/enfermidade/index.blade.php
@@ -7,7 +7,7 @@
         <div class="card">
             <div class="card-body">
                 <a type="button" href="/enfermidade.create" class="btn btn-primary">Incluir</a>
-                
+
                 <table id="dataTable" class="table table-stripped">
                     <thead>
                     <tr>
@@ -19,14 +19,14 @@
                     </thead>
                     <tbody>
                     @foreach($lista as $obj)
-                    
+
                     <input type="hidden" id="id_{{$obj->id}}" value="{{$obj->id}}">
                     <tr>
                         <td width="1%">
                             <a href="/enfermidade.edit.{{$obj->id}}"><i class="fa fa-edit"></i></a>
                         </td>
                         <td width="1%">
-                            <a href="/enfermidade.delete.{{$obj->id}}""><i class="fa fa-trash"></i></a>
+                            <a href="/enfermidade.delete.{{$obj->id}}" onclick="return confirm('Tem certeza de que deseja excluir este item?');"><i class="fa fa-trash"></i></a>
                         </td>
                         <td>
                             {{$obj->descricao}}
