+++ resources/views/paciente/index.blade.php
@@ -20,7 +20,7 @@
         <a href="/paciente.edit.{{ $obj->id }}"><i class="fa fa-edit"></i></a>
       </td>
       <td width="1%">
-        <a href="/paciente.delete.{{ $obj->id }}"><i class="fa fa-trash"></i></a>
+        <a href="/paciente.delete.{{ $obj->id }}" onclick="return confirm('Tem certeza de que deseja excluir este item?');"><i class="fa fa-trash"></i></a>
       </td>
       <td>
         {{ $obj->nome }}
