+++ resources/views/pessoa/index.blade.php
@@ -32,7 +32,7 @@
                             <a href="/pessoa.edit.{{$obj->id}}"><i class="fa fa-edit"></i></a>
                         </td>
                         <td width="1%">
-                            <a href="/pessoa.delete.{{$obj->id}}"><i class="fa fa-trash"></i></a>
+                            <a href="/pessoa.delete.{{$obj->id}}" onclick="return confirm('Tem certeza de que deseja excluir este item?');"><i class="fa fa-trash"></i></a>
                         </td>
                         <td>
                             {{$obj->nome}}
